shopping/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp>=3.8.0
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
auth0-python==4.9.0
aztp-client==1.0.23
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
crewai>=0.120.1
crewai-tools==0.13.2
cryptography>=3.4.0
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.9
durationpy==0.9
embedchain>=0.1.128
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.9
fastavro==1.10.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.3.2
google-auth==2.39.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
gptcache==0.1.44
griffe==1.7.3
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
instructor>=1.7.9
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
json5==0.12.0
json_repair==0.42.0
jsonpatch==1.33
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kubernetes==32.0.1
lancedb==0.22.0
langchain==0.3.23
langchain-cohere==0.3.5
langchain-community==0.3.21
langchain-core==0.3.56
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
langsmith==0.3.34
litellm>=1.30.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.94
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.21.1
openai>=1.12.0,<2.0.0
openai-agents==0.0.2
openpyxl==3.1.5
# opentelemetry-api==1.32.1
# opentelemetry-exporter-otlp-proto-common==1.32.1
# opentelemetry-exporter-otlp-proto-grpc==1.32.1
# opentelemetry-exporter-otlp-proto-http==1.32.1
# opentelemetry-instrumentation==0.53b1
# opentelemetry-instrumentation-asgi==0.53b1
# opentelemetry-instrumentation-fastapi==0.53b1
# opentelemetry-proto==1.32.1
# opentelemetry-sdk==1.32.1
# opentelemetry-semantic-conventions==0.53b1
# opentelemetry-util-http==0.53b1
orjson==3.10.16
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
paypal-agent-toolkit==1.3.0
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
pluggy==1.5.0
portalocker==2.10.1
posthog==3.25.0
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==5.29.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic>=2.0.0
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.4.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.400
pysbd==0.3.4
PySocks==1.7.1
pytest>=6.0.0
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
qdrant-client==1.14.2
referencing==0.36.2
regex==2024.11.6
requests>=2.25.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
schema==0.7.7
selenium==4.31.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.45.3
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.0.1
tiktoken==0.9.0
tokenizers==0.20.3
tomli==2.2.1
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typer==0.15.2
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uv==0.6.16
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
pytest-asyncio>=0.14.0
pytest-cov>=2.10.0
